+++ packages/js-sdk/src/index.ts
@@ -1,4 +1,5 @@
 export { inject } from './inject';
 export { track } from './track';
 export { pageView } from './session';
-export { endSession } from './session'; 
+export { endSession } from './session';
+export { trackClick } from './session';