+++ packages/js-sdk/CHANGELOG.md
@@ -1,5 +1,12 @@
 # lershi-analytics
 
+## 0.3.0
+
+### Minor Changes
+
+- added pageview and endsession
+- added trackclick abstraction with apiurl added
+
 ## 0.2.0
 
 ### Minor Changes