+++ packages/next/CHANGELOG.md
@@ -1,5 +1,18 @@
 # lershi-analytics-next
 
+## 0.3.0
+
+### Minor Changes
+
+- added pageview and endsession
+- added trackclick abstraction with apiurl added
+
+### Patch Changes
+
+- Updated dependencies
+- Updated dependencies
+  - lershi-analytics@0.3.0
+
 ## 0.2.0
 
 ### Minor Changes